Destrinchada classe GameWord em arquivo separado
