+++ index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="stylesheet" href="./styles/main.css" />
-    <script src="./src/index.js" defer></script>
+    <script src="./src/index.js" type="module" defer></script>
     <title>Jogo da Forca</title>
   </head>
   <body>
