+++ src/index.js
@@ -1,20 +1,4 @@
-class GameWord {
-  #word;
-  #category;
-
-  constructor(word, category) {
-    this.#word = word;
-    this.#category = category;
-  }
-
-  get word() {
-    return this.#word;
-  }
-
-  get category() {
-    return this.#category;
-  }
-}
+import { GameWord } from "./GameWord.js";
 
 const wordsList = [
   {
